pyproject.toml
[project]
name = "auto-uploader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "httpx>=0.28.1",
    "pandas>=2.2.3",
    "persiantools>=5.2.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.11",
    "xlsxwriter>=3.2.3",
]

[tool.ruff]
line-length = 88
target-version = "py313"
fix = true
unsafe-fixes = true
preview = true

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"
